Add files via upload
This is a complete Excel project where I worked on the dataset of bike sales. I cleaned the dataset, created pivot tables that summarizes and describes very important insights and I created a final Interactive dashboard report containing charts and visuals that highlights a visual description of the hidden patterns in the dataset.